🚀 refactor(lotto): 로또 출력 메서드 resultView로 이동
