+++ src/main/java/lotto/view/LottoInputView.java
@@ -17,7 +17,6 @@ public class LottoInputView {
     private static final String INPUT_PRICE_MESSAGE = "구입금액을 입력해 주세요.";
     private static final String INPUT_LAST_WEEK_WINNING_NUMBER_MESSAGE = "지난 주 당첨 번호를 입력해 주세요.";
     private static final String INPUT_BONUS_BALL_MESSAGE = "보너스 볼을 입력해주세요.";
-    private static final String LOTTO_AMOUNT_MESSAGE = "개를 구매했습니다.";
 
     public int inputPrice() {
         System.out.println(INPUT_PRICE_MESSAGE);
@@ -40,25 +39,6 @@ public class LottoInputView {
     public LottoNumber inputBonusBall() {
         System.out.println(INPUT_BONUS_BALL_MESSAGE);
         return new LottoNumber(Integer.parseInt(SCANNER.nextLine()));
-    }
-
-    public void printLottos(final Lottos lottos) {
-        printAmount(lottos);
-        for (Lotto lotto : lottos.getLottos()) {
-            printLotto(lotto.getLottoNumbers());
-        }
-    }
-
-    private void printAmount(final Lottos lottos) {
-        System.out.println(lottos.getLottos().size() + LOTTO_AMOUNT_MESSAGE);
-    }
-
-    private void printLotto(final List<LottoNumber> lottoNumbers) {
-        final List<Integer> numbers = lottoNumbers.stream()
-            .map(LottoNumber::getNumber)
-            .collect(Collectors.toList());
-
-        System.out.println(numbers);
     }
 
 }
